requirements.txt
# Core AI & Agent
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20

# Telegram & APIs
python-telegram-bot>=20.7
gspread>=6.0.0
google-auth>=2.23.4
google-api-python-client>=2.111.0  # For Google Calendar API

# Groq API (instead of OpenAI)
groq>=0.4.1
langchain-groq>=0.1.0

# Vector Processing & Search
sentence-transformers>=2.2.2
numpy>=1.24.3

# Task Management
APScheduler>=3.10.4
python-dateutil>=2.8.2

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
lru-dict>=1.2.0
nest-asyncio>=1.6.0
pytz>=2024.1
requests>=2.31.0

# Google Keep (unofficial API)
gkeepapi>=0.16.0